refactor(Restaurant): renamed restaurants.ts and fix en.json, ar.json
