+++ messages/en.json
@@ -400,48 +400,5 @@
   "restaurants.form.eBillSettingsLabel": "E-Bill Settings",
   "restaurants.form.eBillSettings.onEmail": "Email E-Bills",
   "restaurants.form.eBillSettings.onWhatsapp": "WhatsApp E-Bills",
-  "restaurants.form.eBillSettings.onSms": "SMS E-Bills",
-
-  "locations.title": "Location Management",
-  "locations.subtitle": "Manage restaurant locations and delivery areas",
-  "locations.addNewLocation": "Add New Location",
-  "locations.allLocations": "All Locations",
-  "locations.searchPlaceholder": "Search locations...",
-  "locations.restaurant": "Restaurant",
-  "locations.area": "Area",
-  "locations.deliveryZone": "Delivery Zone",
-  "locations.coordinates": "Coordinates",
-  "locations.status": "Status",
-  "locations.active": "Active",
-  "locations.inactive": "Inactive",
-  "locations.edit": "Edit",
-  "locations.delete": "Delete",
-  "locations.view": "View on Map",
-
-  "locations.form.createTitle": "Create New Location",
-  "locations.form.editTitle": "Edit Location",
-  "locations.form.createDescription": "Fill in the information below to create a new restaurant location.",
-  "locations.form.editDescription": "Update the location information below.",
-  "locations.form.basicInfo": "Basic Information",
-  "locations.form.restaurantLabel": "Restaurant",
-  "locations.form.restaurantPlaceholder": "Select a restaurant",
-  "locations.form.areaLabel": "Area",
-  "locations.form.areaPlaceholder": "Enter area (e.g., Downtown, Mall)",
-  "locations.form.areaDescription": "The general area of the restaurant location.",
-  "locations.form.deliveryZoneLabel": "Delivery Zone",
-  "locations.form.deliveryZonePlaceholder": "Enter delivery zone name",
-  "locations.form.deliveryZoneDescription": "The specific delivery zone covered by this location.",
-  "locations.form.activeStatusLabel": "Active Status",
-  "locations.form.activeStatusDescription": "Toggle to activate or deactivate this location.",
-  "locations.form.deliveryInfo": "Delivery Information",
-  "locations.form.deliveryRadiusLabel": "Delivery Radius (km)",
-  "locations.form.deliveryRadiusPlaceholder": "Enter delivery radius in kilometers",
-  "locations.form.deliveryRadiusDescription": "The maximum distance for deliveries from this location.",
-  "locations.form.coordinatesInfo": "Coordinates Information",
-  "locations.form.latitudeLabel": "Latitude",
-  "locations.form.latitudePlaceholder": "Enter latitude",
-  "locations.form.latitudeDescription": "The latitude of the restaurant location.",
-  "locations.form.longitudeLabel": "Longitude",
-  "locations.form.longitudePlaceholder": "Enter longitude",
-  "locations.form.longitudeDescription": "The longitude of the restaurant location."
+  "restaurants.form.eBillSettings.onSms": "SMS E-Bills"
 }
